task2/requirements.txt
numpy==1.21.6
Pillow==10.0.0
PyQt5==5.15.9
PyQt5_sip==12.12.1
scikit_learn==1.0.2
selenium==4.10.0
